Alternative solution to expose embeded fields
